"move 'prepare_test_data()' and 'excel_to_matrix' to tests.py"
